# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Marketing Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{clientId}/memory:
    get:
      operationId: listMemory
      tags: [content-memory]
      summary: List content memory for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    post:
      operationId: addMemory
      tags: [content-memory]
      summary: Add a memory entry
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemoryBody"
      responses:
        "201":
          description: Memory entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /clients/{clientId}/memory/{memoryId}:
    delete:
      operationId: deleteMemory
      tags: [content-memory]
      summary: Delete a memory entry
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── AI Generation ─────────────────────────────────────────────────────────
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                required: [accessToken, refreshToken]

  # ── Settings ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MemoryEntry:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        key:
          type: string
        value:
          type: string
        createdAt:
          type: string
      required: [id, clientId, key, value, createdAt]
    AddMemoryBody:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required: [key, value]

    # ── AI ─────────────────────────────────────────────────────────────────────
...